// openvidu-call-front/src/app/shared/types/singal-type.ts
export enum Signal {
	NICKNAME_CHANGED = 'nicknameChanged',
	DISABLE_AUDIO = 'disableAudio',
	ENABLE_AUDIO = 'enableAudio',
	TOGGLE_VIDEO = 'toggleVideo',
	MESSAGE_TYPING='messageTyping',
	WHITEBOARD_CLEAR = 'whiteboardClear',
	WHITEBOARD_REDO = 'whiteboardRedo',
	WHITEBOARD_UNDO = 'whiteboardUndo',
	WHITEBOARD_DRAW = 'whiteboardDraw',
	WHITEBOARD_TOGGLE = 'whiteboardToggle'

}